chore: move compatibility layer to fileutils package
