--- pkg/management/postgres/compatibility/windows.go
@@ -1,34 +1,0 @@
-//go:build windows
-// +build windows
-
-/*
-Copyright The CloudNativePG Contributors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package compatibility
-
-import (
-	"os/exec"
-)
-
-// AddInstanceRunCommands adds specific OS commands to the postgres exec.Cmd
-func AddInstanceRunCommands(cmd *exec.Cmd) {
-	return
-}
-
-// Umask sets the process's unix umask to prevent/allow permissions changes
-func Umask(mask int) int {
-	return mask
-}
